requirements.txt
openai==0.27.2
line-bot-sdk==2.4.2
